% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffStyle.R
\name{AffStyle}
\alias{AffStyle}
\title{Automatic generation of author list with different styles.}
\usage{
AffStyle(name, degree, index, style)
}
\arguments{
\item{name}{Authors' names}

\item{degree}{Authors' title (PhD, MS, etc..)}

\item{index}{\code{index} acquired from \code{AutoAff} function. See 'Value' in \link{AutoAff}.}

\item{style}{Style of output. See 'Details'.}
}
\value{
The function returns a text string that is either ready to be pasted into your manuscript (without superscripts) or need to be copied to R Markdown and knitted (with superscripts).
}
\description{
\code{AffStyle} facilitates the generation of long author list in different styles based on the requirements of research journals.
}
\details{
Currently supporting two styles but growing.

plain: John Doe, MSc 1-3; Jane Doe, PhD 1,3-5; ..

superscript: John Doe, MSc^1-3^; Jane Doe, PhD^1,3-5^; ..

Styles that require superscripted indices require the use of R Markdown. The output for these styles are R Markdown syntax (text between ^^ will be superscipts after knitting.)
}
\examples{
result <- AutoAff(authors, c("aff1", "aff2", "aff3", "aff4", "aff5", "aff6"))
AffStyle(name = authors$Name, degree = authors$Degree, index = result$index, style = "plain")

}
