% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\docType{data}
\name{authors}
\alias{authors}
\title{A demo dataset}
\format{A data frame with 100 rows and 7 variables:
\describe{
  \item{Name}{Author's name}
  \item{aff1}{First affiliation}
  \item{aff2}{Second affiliation}
  ...
  \item{aff6}{Sixth affiliation}
}}
\usage{
data(authors)
}
\description{
A fake data frame containing 100 authors with up to 6 affiliations.
}
\keyword{datasets}
