% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoAff.R
\name{AutoAff}
\alias{AutoAff}
\title{Automatic generation of affiliation lists.}
\usage{
AutoAff(X, affiliation)
}
\arguments{
\item{X}{A user-supplied data frame with columns containing the authors' affiliations.}

\item{affiliation}{Names of the variables containing affiliations. A character vector, see 'Examples'.}
}
\value{
This function returns a list with six elements:
\item{rmd}{Indexed affiliations in R Markdown syntax, see 'Details'.}
\item{affiliations}{Affiliations by order of appearance}
\item{aff.count}{Number of different affiliations}
\item{aut.count}{Number of authors}
\item{index}{Indices of each author's affililation(s)}
\item{missing}{Number of authors with no affiliations}
}
\description{
\code{AutoAff} facilitates the management for long author lists and their affiliations. The outputs include a text string of R Markdown syntax that can be knitted to create superscripted indices.
}
\details{
Empty cells (for those authors with fewer affiliations) in \code{X} have to be either \code{NA} or \code{""}.

Copy and paste \code{rmd} in \code{result} into R Markdown and knit to Microsoft Word to get affiliation lists with superscripted indices. See 'Examples'.
}
\examples{
result <- AutoAff(X = authors,
                  affiliation = c("aff1", "aff2", "aff3", "aff4", "aff5", "aff6"))
result

# R Markdown syntax
result$rmd

}
