% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffStyle.R
\name{AffStyle}
\alias{AffStyle}
\title{Automatic generation of author list with customized styles.}
\usage{
AffStyle(X, first, last, middle, name, degree, index, template)
}
\arguments{
\item{X}{A user-supplied data frame with columns containing the authors' affiliations.}

\item{first}{Name of the variable containing the authors' first names. See 'Examples'.}

\item{last}{Name of the variable containing the authors' last names. See 'Examples'.}

\item{middle}{Name of the variable containing the authors' middle names. See 'Examples'.}

\item{name}{Name of the variable containing the authors' full names. See 'Examples'.}

\item{degree}{Name of the variable containing the authors' title (PhD, MS, etc..). See 'Examples'.}

\item{index}{\code{index} acquired from \code{AutoAff} function. See 'Value' in \link{AutoAff}.}

\item{template}{A template indicating the style of output. See 'Details' and 'Examples'.}
}
\value{
The function returns a text string that is either ready to be pasted into your manuscript (without superscripts) or need to be copied to R Markdown and knitted (with superscripts).
}
\description{
\code{AffStyle} facilitates the generation of long author list in different styles based on the requirements of research journals.
}
\details{
\code{template} represents a style specified by the user.

f: first name

m: middle name

l: last name

n: full name

d: degree

i: index (acquired by \code{AutoAff})

^foo^: if 'foo' needs to be superscripted

Any other characters can be used based on the user's need. For example, spaces can be added between first and last names, colons ';' can be added at the end as separators.

Styles that include superscripts/subscripts require the use of R Markdown. The output for these styles are R Markdown syntax (text between ^^ will be superscipts after knitting).
}
\examples{
result <- AutoAff(X = authors,
                  affiliation = c("aff1", "aff2", "aff3", "aff4", "aff5", "aff6"))#' 
template <- "f m l, d^i^;"
AffStyle(X = authors, first = "First", middle = "Middle", last = "Last",
         degree = "Degree", index = result$index, template = template)

}
